Refactor stylesheet_url with better tools and remove useless code.
